🚚 chore: Rename HeaderBar folder to Header
