+++ frontend/src/App.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Header from './components/HeaderBar/Header';
+import Header from './components/Header/Header';
 
 const App = () => {
   return (
